Refactored Nation enum into separate file
