+++ cwms-radar-client/src/main/java/mil/army/usace/hec/cwms/radar/client/model/Location.java
@@ -1,8 +1,6 @@
 package mil.army.usace.hec.cwms.radar.client.model;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonValue;
 import io.swagger.annotations.ApiModelProperty;
 
 import java.util.Objects;
@@ -364,38 +362,5 @@ public class Location {
             return "null";
         }
         return o.toString().replace("\n", "\n    ");
-    }
-
-    /**
-     * Gets or Sets nation
-     */
-    public enum Nation {
-        US("US"),
-
-        CANADA("CANADA"),
-
-        MEXICO("MEXICO");
-
-        private String value;
-
-        Nation(String value) {
-            this.value = value;
-        }
-
-        @JsonCreator
-        public static Nation fromValue(String text) {
-            for (Nation b : Nation.values()) {
-                if (String.valueOf(b.value).equals(text)) {
-                    return b;
-                }
-            }
-            return null;
-        }
-
-        @Override
-        @JsonValue
-        public String toString() {
-            return String.valueOf(value);
-        }
     }
 }
